Adds axis correction, separates buildDrawCommand
