+++ Source/Scene/ModelExperimentalSceneGraph.js
@@ -3,13 +3,10 @@ import Matrix4 from "../Core/Matrix4.js";
 import ModelExperimentalSceneMeshPrimitive from "./ModelExperimentalSceneMeshPrimitive.js";
 import ModelExperimentalSceneNode from "./ModelExperimentalSceneNode.js";
 import ModelExperimentalUtility from "./ModelExperimentalUtility.js";
-import VertexArray from "../Renderer/VertexArray.js";
-import DrawCommand from "../Renderer/DrawCommand.js";
+import buildDrawCommand from "./buildDrawCommand.js";
 import defaultValue from "../Core/defaultValue.js";
-import Pass from "../Renderer/Pass.js";
 import RenderResources from "./RenderResources.js";
 import Axis from "../Scene/Axis.js";
-import RenderState from "../Renderer/RenderState.js";
 
 export default function ModelExperimentalSceneGraph(options) {
   /**
@@ -45,6 +42,12 @@ export default function ModelExperimentalSceneGraph(options) {
 
 function initialize(sceneGraph) {
   var modelMatrix = sceneGraph._model.modelMatrix;
+
+  ModelExperimentalUtility.correctModelMatrix(
+    modelMatrix,
+    sceneGraph._upAxis,
+    sceneGraph._forwardAxis
+  );
 
   var rootNode = sceneGraph._modelComponents.scene.nodes[0];
   var rootNodeModelMatrix = Matrix4.multiply(
@@ -151,56 +154,3 @@ ModelExperimentalSceneGraph.prototype.buildDrawCommands = function (
     }
   }
 };
-
-/**
- * Builds a DrawCommand for a glTF mesh primitive using its render resources.
- *
- * @param {RenderResources.MeshRenderResources} meshPrimitiveRenderResources
- * @param {FrameState} frameState
- * @returns {DrawCommand} The generated DrawCommand.
- */
-function buildDrawCommand(meshPrimitiveRenderResources, frameState) {
-  var shaderBuilder = meshPrimitiveRenderResources.shaderBuilder;
-  shaderBuilder.addVertexLines([
-    "void main()",
-    "{",
-    "    vec3 position = a_position;",
-    "    gl_Position = czm_modelViewProjection * vec4(position, 1.0);",
-    "}",
-  ]);
-  shaderBuilder.addFragmentLines([
-    "void main()",
-    "{",
-    "    vec4 color = vec4(0.0, 0.0, 1.0, 1.0);",
-    "    gl_FragColor = color;",
-    "}",
-  ]);
-
-  var vertexArray = new VertexArray({
-    context: frameState.context,
-    indexBuffer: meshPrimitiveRenderResources.indices.buffer,
-    attributes: meshPrimitiveRenderResources.attributes,
-  });
-
-  var renderState = RenderState.fromCache(
-    meshPrimitiveRenderResources.renderStateOptions
-  );
-
-  var shaderProgram = shaderBuilder.buildShaderProgram(frameState.context);
-
-  return new DrawCommand({
-    boundingVolume: meshPrimitiveRenderResources.boundingSphere,
-    modelMatrix: meshPrimitiveRenderResources.modelMatrix,
-    uniformMap: meshPrimitiveRenderResources.uniformMap,
-    renderState: renderState,
-    vertexArray: vertexArray,
-    shaderProgram: shaderProgram,
-    cull: false,
-    pass: Pass.OPAQUE,
-    count: meshPrimitiveRenderResources.count,
-    pickId: undefined,
-    instanceCount: 0,
-    primitiveType: meshPrimitiveRenderResources.primitiveType,
-    debugShowBoundingVolume: true,
-  });
-}
